Spark job mixin created
