+++ edx/analytics/tasks/monitor/overall_events.py
@@ -5,9 +5,9 @@ import logging
 import luigi
 
 from edx.analytics.tasks.common.mapreduce import MapReduceJobTask
-from edx.analytics.tasks.common.pathutil import EventLogSelectionMixin, EventLogSelectionMixinSpark
+from edx.analytics.tasks.common.pathutil import EventLogSelectionMixin
+from edx.analytics.tasks.common.spark import EventLogSelectionMixinSpark, SparkJobTask
 from edx.analytics.tasks.util.url import get_target_from_url
-from luigi.contrib.spark import PySparkTask
 
 log = logging.getLogger(__name__)
 
@@ -37,7 +37,7 @@ class TotalEventsDailyTask(EventLogSelectionMixin, MapReduceJobTask):
         return get_target_from_url(self.output_root)
 
 
-class SparkTotalEventsDailyTask(EventLogSelectionMixinSpark, PySparkTask):
+class SparkTotalEventsDailyTask(EventLogSelectionMixinSpark, SparkJobTask):
     """Produce a dataset for total events within a given time period."""
 
     driver_memory = '2g'
@@ -45,46 +45,12 @@ class SparkTotalEventsDailyTask(EventLogSelectionMixinSpark, PySparkTask):
 
     output_root = luigi.Parameter()
 
-    def __init__(self, *args, **kwargs):
-        super(SparkTotalEventsDailyTask, self).__init__(*args, **kwargs)
-
     def output(self):
         return get_target_from_url(self.output_root)
 
-    def main(self, sc, *args):
-        from pyspark.sql import SparkSession
-        from pyspark.sql.types import *
+    def spark_job(self):
         from pyspark.sql.functions import to_date, udf, struct, date_format
-        spark = SparkSession.builder.getOrCreate()
-        event_schema = StructType().add("POST", StringType(), True).add("GET", StringType(), True)
-        module_schema = StructType().add("display_name", StringType(), True) \
-            .add("original_usage_key", StringType(), True) \
-            .add("original_usage_version", StringType(), True) \
-            .add("usage_key", StringType(), True)
-        context_schema = StructType().add("command", StringType(), True) \
-            .add("course_id", StringType(), True) \
-            .add("module", module_schema) \
-            .add("org_id", StringType(), True) \
-            .add("path", StringType(), True) \
-            .add("user_id", StringType(), True)
-
-        event_log_schema = StructType() \
-            .add("username", StringType(), True) \
-            .add("event_type", StringType(), True) \
-            .add("ip", StringType(), True) \
-            .add("agent", StringType(), True) \
-            .add("host", StringType(), True) \
-            .add("referer", StringType(), True) \
-            .add("accept_language", StringType(), True) \
-            .add("event", event_schema) \
-            .add("event_source", StringType(), True) \
-            .add("context", context_schema) \
-            .add("time", StringType(), True) \
-            .add("name", StringType(), True) \
-            .add("page", StringType(), True) \
-            .add("session", StringType(), True)
-
-        df = spark.read.format('json').load(self.path_targets, schema=event_log_schema)
+        df = self._spark.read.format('json').load(self.path_targets, schema=self.get_log_schema())
         df = df.filter(df['time'].isNotNull()) \
             .withColumn('event_date', date_format(to_date(df['time']), 'yyyy-MM-dd'))
         df = df.filter(df['event_date'] == self.lower_bound_date_string).groupBy('event_date').count()
